Temporarily disabling orphaned events deletion
Currently we check for ALL events in both main and archived orgmode
file. This is good, as we can update events that ended up in archive
before they were updated in the calendar, however this will take more
and more time to sync as far as we add more and more tasks in orgmode.
To avoid that, I could
1. make deletion optional, or guarded with manual confirm (which is kind
   of boring)
2. remove archive.org file from the watch list

while (2) is much better, the current implementation would delete a
bunch of events that exists in the archive.md.
So, for now, I just skip the deletion part

+++ cmd/sync.go
@@ -90,20 +90,29 @@ func sync(calendarName string, tasks []model.Task) {
 		taskMap[task.ID] = true
 	}
 
-	// Fetch recent events to check for orphans
-	events, err := client.ListEvents(time.Now().Add(-30 * 24 * time.Hour))
-	if err != nil {
-		log.Fatalf("Error fetching calendar events: %v", err)
-	}
+	// TODO: improve orphaned events management. For now it is disabled
+	// the problem is that looking for all the events will take more and more
+	// time in the future, since we'll have more and more events in history.
+	// The solution is to look only in Orgmode.md not in Orgmode_archive.md. However,
+	// as of today implementation, removing the latter from the watch list, will make a
+	// bunch of events form the Calendar "orphaned" and so deleted.
 
-	// Delete orphaned events
-	for _, event := range events {
-		taskID, found := util.GetTaskIDFromEventDescription(event.Description)
-		if found && !taskMap[taskID] {
-			log.Printf("Deleting orphaned event for task '%s' ID %s", event.Description, taskID)
-			err := client.DeleteEvent(event.Id)
-			if err != nil {
-				log.Printf("Error deleting event: %v", err)
+	if false {
+		// Fetch recent events to check for orphans
+		events, err := client.ListEvents(time.Now().Add(-30 * 24 * time.Hour))
+		if err != nil {
+			log.Fatalf("Error fetching calendar events: %v", err)
+		}
+
+		// Delete orphaned events
+		for _, event := range events {
+			taskID, found := util.GetTaskIDFromEventDescription(event.Description)
+			if found && !taskMap[taskID] {
+				log.Printf("Deleting orphaned event for task '%s' ID %s", event.Description, taskID)
+				err := client.DeleteEvent(event.Id)
+				if err != nil {
+					log.Printf("Error deleting event: %v", err)
+				}
 			}
 		}
 	}
